@@ -1,12 +1,11 @@
-// ignore_for_file: prefer_const_constructors, deprecated_member_use
-
 import 'package:flutter/material.dart';
 
-class Badge extends StatelessWidget {
+class CustomBadge extends StatelessWidget {
   final Widget child;
   final String value;
   final Color? color;
-  const Badge({
+
+  const CustomBadge({
     Key? key,
     required this.child,
     required this.value,
@@ -26,16 +25,16 @@ class Badge extends StatelessWidget {
             padding: const EdgeInsets.all(2),
             decoration: BoxDecoration(
               borderRadius: BorderRadius.circular(10),
-              color: color ?? Theme.of(context).accentColor,
+              color: color ?? Theme.of(context).colorScheme.secondary,
             ),
-            constraints: BoxConstraints(
+            constraints: const BoxConstraints(
               minHeight: 16,
               minWidth: 16,
             ),
             child: Text(
               value,
               textAlign: TextAlign.center,
-              style: TextStyle(
+              style: const TextStyle(
                 fontSize: 10,
               ),
             ),
