@@ -6,10 +6,7 @@ import 'package:shop/models/cart_item.dart';
 class CartItemWidget extends StatelessWidget {
   final CartItem cartItem;
 
-  const CartItemWidget(
-    this.cartItem, {
-    Key? key,
-  }) : super(key: key);
+  const CartItemWidget(this.cartItem, {Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -18,37 +15,37 @@ class CartItemWidget extends StatelessWidget {
       direction: DismissDirection.endToStart,
       background: Container(
         color: Theme.of(context).errorColor,
-        child: Icon(
-          Icons.delete,
-          color: Colors.white,
-          size: 40,
-        ),
         alignment: Alignment.centerRight,
         padding: const EdgeInsets.only(right: 20),
         margin: const EdgeInsets.symmetric(
           horizontal: 15,
           vertical: 4,
         ),
+        child: const Icon(
+          Icons.delete,
+          color: Colors.white,
+          size: 40,
+        ),
       ),
       confirmDismiss: (_) {
         return showDialog<bool>(
           context: context,
           builder: (ctx) => AlertDialog(
-            title: const Text('Tem Certeza ?'),
-            content: const Text('Quer remover o item do carrinho ?'),
+            title: const Text('Tem Certeza?'),
+            content: const Text('Quer remover o item do carrinho?'),
             actions: [
               TextButton(
+                child: const Text('Não'),
                 onPressed: () {
                   Navigator.of(ctx).pop(false);
                 },
-                child: const Text('Não'),
               ),
               TextButton(
+                child: const Text('Sim'),
                 onPressed: () {
                   Navigator.of(ctx).pop(true);
                 },
-                child: const Text('Sim'),
-              )
+              ),
             ],
           ),
         );
@@ -68,6 +65,8 @@ class CartItemWidget extends StatelessWidget {
           padding: const EdgeInsets.all(8),
           child: ListTile(
             leading: CircleAvatar(
+              backgroundColor: Theme.of(context).colorScheme.primary,
+              foregroundColor: Colors.white,
               child: Padding(
                 padding: const EdgeInsets.all(5),
                 child: FittedBox(
