@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-
-import '../models/order.dart';
+import 'package:shop/models/order.dart';
 
 class OrderWidget extends StatefulWidget {
   final Order order;
+
   const OrderWidget({
     Key? key,
     required this.order,
@@ -16,6 +16,7 @@ class OrderWidget extends StatefulWidget {
 
 class _OrderWidgetState extends State<OrderWidget> {
   bool _expanded = false;
+
   @override
   Widget build(BuildContext context) {
     return Card(
@@ -27,9 +28,7 @@ class _OrderWidgetState extends State<OrderWidget> {
               DateFormat('dd/MM/yyyy hh:mm').format(widget.order.date),
             ),
             trailing: IconButton(
-              icon: const Icon(
-                Icons.expand_more,
-              ),
+              icon: const Icon(Icons.expand_more),
               onPressed: () {
                 setState(() {
                   _expanded = !_expanded;
@@ -43,7 +42,7 @@ class _OrderWidgetState extends State<OrderWidget> {
                 horizontal: 15,
                 vertical: 4,
               ),
-              height: (widget.order.products.length * 25.0) + 10,
+              height: (widget.order.products.length * 24) + 10,
               child: ListView(
                 children: widget.order.products.map(
                   (product) {
