@@ -1,4 +1,4 @@
-import 'cart_item.dart';
+import 'package:shop/models/cart_item.dart';
 
 class Order {
   final String id;
@@ -6,9 +6,10 @@ class Order {
   final List<CartItem> products;
   final DateTime date;
 
-  Order(
-      {required this.id,
-      required this.total,
-      required this.products,
-      required this.date});
+  Order({
+    required this.id,
+    required this.total,
+    required this.products,
+    required this.date,
+  });
 }
