@@ -1,10 +1,11 @@
-import 'dart:math';
+import 'dart:convert';
+import 'package:flutter/material.dart';
+import 'package:shop/models/cart.dart';
+import 'package:shop/models/cart_item.dart';
+import 'package:shop/models/order.dart';
+import 'package:http/http.dart' as http;
 
-import 'package:flutter/widgets.dart';
-
-import 'cart.dart';
-import 'cart_item.dart';
-import 'order.dart';
+import '../utils/constants.dart';
 
 class OrderList with ChangeNotifier {
   final List<Order> _items = [];
@@ -17,13 +18,65 @@ class OrderList with ChangeNotifier {
     return _items.length;
   }
 
-  void addOrder(Cart cart) {
+  Future<void> loadOrders() async {
+    _items.clear();
+
+    final response = await http.get(
+      Uri.parse('${Constants.ORDER_BASE_URL}.json'),
+    );
+    if (response.body == 'null') return;
+    Map<String, dynamic> data = jsonDecode(response.body);
+    data.forEach((orderId, orderData) {
+      _items.add(Order(
+        id: orderId,
+        date: DateTime.parse(orderData['date']),
+        total: orderData['total'],
+        products: orderData['products']
+            .map((item) => {
+                  CartItem(
+                    id: item['id'],
+                    productId: item['productId'],
+                    name: item['name'],
+                    quantity: item['quantity'],
+                    price: item['price'],
+                  )
+                })
+            .toList(),
+      ));
+    });
+    notifyListeners();
+  }
+
+  Future<void> addOrder(Cart cart) async {
+    final date = DateTime.now();
+
+    final response = await http.post(
+      Uri.parse('${Constants.ORDER_BASE_URL}.json'),
+      body: jsonEncode(
+        {
+          'total': cart.totalAmount,
+          'date': date.toIso8601String(),
+          'products': cart.items.values
+              .map((cartItem) => {
+                    'id': cartItem.id,
+                    'productId': cartItem.productId,
+                    'name': cartItem.name,
+                    'quantity': cartItem.quantity,
+                    'price': cartItem.price,
+                  })
+              .toList(),
+        },
+      ),
+    );
+
+    final id = jsonDecode(response.body)['name'];
+
     _items.insert(
       0,
       Order(
-        id: Random().nextDouble().toString(),
+        id: id,
         total: cart.totalAmount,
-        date: DateTime.now(),
+        date: date,
         products: cart.items.values.toList(),
       ),
     );
