@@ -1,4 +1,9 @@
-import 'package:flutter/widgets.dart';
+import 'dart:convert';
+
+import 'package:flutter/material.dart';
+import 'package:http/http.dart' as http;
+
+import '../utils/constants.dart';
 
 class Product with ChangeNotifier {
   final String id;
@@ -17,8 +22,29 @@ class Product with ChangeNotifier {
     this.isFavorite = false,
   });
 
-  void toggleFavorite() {
+  void _toggleFavorite() {
     isFavorite = !isFavorite;
     notifyListeners();
+  }
+
+  Future<void> toggleFavorite() async {
+    try {
+      _toggleFavorite();
+
+      final response = await http.patch(
+        Uri.parse('${Constants.PRODUCT_BASE_URL}/$id.json'),
+        body: jsonEncode(
+          {
+            "isFavorite": isFavorite,
+          },
+        ),
+      );
+
+      if (response.statusCode >= 400) {
+        _toggleFavorite();
+      }
+    } catch (_) {
+      _toggleFavorite();
+    }
   }
 }
