@@ -2,9 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:shop/providers/counter.dart';
 
 class CounterPage extends StatefulWidget {
-  const CounterPage({
-    Key? key,
-  }) : super(key: key);
+  const CounterPage({Key? key}) : super(key: key);
 
   @override
   State<CounterPage> createState() => _CounterPageState();
@@ -13,30 +11,30 @@ class CounterPage extends StatefulWidget {
 class _CounterPageState extends State<CounterPage> {
   @override
   Widget build(BuildContext context) {
+    final provider = CounterProvider.of(context);
+
     return Scaffold(
       appBar: AppBar(
         title: const Text('Exemplo Contador'),
       ),
       body: Column(
         children: [
-          Text(CounterProvider.of(context)?.state.value.toString() ?? '0'),
+          Text(provider?.state.value.toString() ?? '0'),
           IconButton(
             onPressed: () {
               setState(() {
-                CounterProvider.of(context)?.state.inc();
+                provider?.state.inc();
               });
-              print(CounterProvider.of(context)?.state.value.toString());
+              // print(provider?.state.value);
             },
-            icon: const Icon(
-              Icons.add,
-            ),
+            icon: const Icon(Icons.add),
           ),
           IconButton(
             onPressed: () {
               setState(() {
-                CounterProvider.of(context)?.state.dec();
+                provider?.state.dec();
               });
-              print(CounterProvider.of(context)?.state.value.toString());
+              // print(provider?.state.value);
             },
             icon: const Icon(Icons.remove),
           ),
