@@ -1,29 +1,46 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-
-import '../components/app_drawer.dart';
-import '../components/order.dart';
-import '../models/order_list.dart';
+import 'package:shop/components/app_drawer.dart';
+import 'package:shop/components/order.dart';
+import 'package:shop/models/order_list.dart';
 
 class OrdersPage extends StatelessWidget {
   const OrdersPage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    final OrderList orders = Provider.of(context);
-
     return Scaffold(
       appBar: AppBar(
-        title: Text('Meus pedidos'),
+        title: const Text('Meus Pedidos'),
       ),
-      drawer: AppDrawer(),
-      body: ListView.builder(
-        itemCount: orders.itemsCount,
-        itemBuilder: (ctx, i) => OrderWidget(
-          order: orders.items[i],
-        ),
+      drawer: const AppDrawer(),
+      body: FutureBuilder(
+        future: Provider.of<OrderList>(context, listen: false).loadOrders(),
+        builder: (ctx, snapshot) {
+          if (snapshot.connectionState == ConnectionState.waiting) {
+            return const Center(child: CircularProgressIndicator());
+          } else if (snapshot.error != null) {
+            return const Center(
+              child: Text('Ocorreu um erro! '),
+            );
+          } else {
+            return Consumer<OrderList>(
+              builder: (ctx, orders, child) => ListView.builder(
+                itemCount: orders.itemsCount,
+                itemBuilder: (ctx, i) => OrderWidget(
+                  order: orders.items[i],
+                ),
+              ),
+            );
+          }
+        },
       ),
+      // body: _isLoading
+      //     ? const Center(child: CircularProgressIndicator())
+      //     : ListView.builder(
+      //         itemCount: orders.itemsCount,
+      //         itemBuilder: (ctx, i) => OrderWidget(order: orders.items[i]),
+      //       ),
     );
   }
 }
