@@ -1,17 +1,13 @@
 import 'package:flutter/material.dart';
-
-import '../models/product.dart';
+import 'package:shop/models/product.dart';
 
 class ProductDetailPage extends StatelessWidget {
-  ProductDetailPage({
-    Key? key,
-  }) : super(key: key);
+  const ProductDetailPage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     final Product product =
         ModalRoute.of(context)!.settings.arguments as Product;
-
     return Scaffold(
       appBar: AppBar(
         title: Text(product.name),
@@ -19,7 +15,7 @@ class ProductDetailPage extends StatelessWidget {
       body: SingleChildScrollView(
         child: Column(
           children: [
-            Container(
+            SizedBox(
               height: 300,
               width: double.infinity,
               child: Image.network(
@@ -27,17 +23,17 @@ class ProductDetailPage extends StatelessWidget {
                 fit: BoxFit.cover,
               ),
             ),
-            SizedBox(height: 10),
+            const SizedBox(height: 10),
             Text(
               'R\$ ${product.price}',
-              style: TextStyle(
-                fontSize: 20,
+              style: const TextStyle(
                 color: Colors.grey,
+                fontSize: 20,
               ),
             ),
-            SizedBox(height: 10),
+            const SizedBox(height: 10),
             Container(
-              padding: EdgeInsets.all(10),
+              padding: const EdgeInsets.symmetric(horizontal: 10),
               width: double.infinity,
               child: Text(
                 product.description,
