@@ -1,7 +1,8 @@
-import 'package:flutter/widgets.dart';
+import 'package:flutter/material.dart';
 
 class CounterState {
   int _value = 0;
+
   void inc() => _value++;
   void dec() => _value--;
   int get value => _value;
@@ -14,10 +15,8 @@ class CounterState {
 class CounterProvider extends InheritedWidget {
   final CounterState state = CounterState();
 
-  CounterProvider({
-    Key? key,
-    required Widget child,
-  }) : super(key: key, child: child);
+  CounterProvider({Key? key, required Widget child})
+      : super(key: key, child: child);
 
   static CounterProvider? of(BuildContext context) {
     return context.dependOnInheritedWidgetOfExactType<CounterProvider>();
@@ -25,6 +24,6 @@ class CounterProvider extends InheritedWidget {
 
   @override
   bool updateShouldNotify(covariant CounterProvider oldWidget) {
-    return oldWidget.state.diff(state);
+    throw oldWidget.state.diff(state);
   }
 }
